+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Spielrunden werden nummeriert
 * PriceTags werden bei der Erzeugung der Preisliste berechnet (für Darstellung Karte)
 * BREAKING: Team Members werden mit Name gespeichert -> Anpassung Teams Schema
+* Anpassungen Logs
+* Speed-Demo Game mit maximalen Settings
+* Bugfix bei Notification Main
 
 # v4.1.0 25.9.25
 * Verlorenes Feature nachgerüstet: Update auf Filter "Verwendung" in Editor kann auf Orte in Preisliste filtern